+++ lib/streamlit/delta_generator.py
@@ -21,7 +21,7 @@ from streamlit import legacy_caching
 from streamlit import type_util
 from streamlit import util
 from streamlit.cursor import Cursor
-from streamlit.script_run_context import get_script_run_ctx
+from streamlit.scriptrunner.script_run_context import get_script_run_ctx
 from streamlit.errors import StreamlitAPIException
 from streamlit.errors import NoSessionContext
 from streamlit.proto import Block_pb2
