+++ lib/streamlit/elements/color_picker.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 
 import re
-from streamlit.script_run_context import ScriptRunContext, get_script_run_ctx
+from streamlit.scriptrunner.script_run_context import ScriptRunContext, get_script_run_ctx
 from streamlit.type_util import Key, to_key
 from textwrap import dedent
 from typing import Optional, cast
