+++ lib/streamlit/scriptrunner/__init__.py
@@ -12,13 +12,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import pytest
-from tests import testutil
-from streamlit.scriptrunner.script_runner import StopException
-import streamlit as st
+# Explicitly export public symobls
+from .script_runner import (
+    ScriptRunner as ScriptRunner,
+    ScriptRunnerEvent as ScriptRunnerEvent,
+)
 
-
-class StopTest(testutil.DeltaGeneratorTestCase):
-    def test_stop(self):
-        with pytest.raises(StopException) as exc_message:
-            st.stop()
+from .script_run_context import (
+    ScriptRunContext as ScriptRunContext,
+    get_script_run_ctx as get_script_run_ctx,
+)
