+++ lib/tests/streamlit/caching/common_cache_test.py
@@ -20,14 +20,14 @@ from unittest.mock import patch
 from parameterized import parameterized
 
 import streamlit as st
-from streamlit import script_run_context
+from streamlit.scriptrunner import script_run_context
 from streamlit.caching import (
     MEMO_CALL_STACK,
     SINGLETON_CALL_STACK,
 )
 from streamlit.forward_msg_queue import ForwardMsgQueue
 
-from streamlit.script_run_context import (
+from streamlit.scriptrunner.script_run_context import (
     add_script_run_ctx,
     get_script_run_ctx,
     ScriptRunContext,
