+++ lib/tests/streamlit/report_context_test.py
@@ -16,7 +16,7 @@ import unittest
 
 from streamlit.errors import StreamlitAPIException
 from streamlit.proto.ForwardMsg_pb2 import ForwardMsg
-from streamlit.script_run_context import ScriptRunContext
+from streamlit.scriptrunner.script_run_context import ScriptRunContext
 from streamlit.state.session_state import SessionState
 from streamlit.uploaded_file_manager import UploadedFileManager
